set seeds with omnet ini file

+++ suqc/environment.py
@@ -8,7 +8,6 @@ import time
 from shutil import copytree, ignore_patterns, rmtree
 from typing import *
 
-import utils
 from suqc.configuration import SuqcConfig
 from suqc.opp.config_parser import OppConfigFileBase, OppConfigType, OppParser
 from suqc.utils.general import (
@@ -324,6 +323,10 @@ class AbstractEnvironmentManager(object):
         output_folder_path = os.path.join(base_path, env_name)
         return output_folder_path
 
+    def write_data(self, path):
+        with open(path, "w+") as f:
+            f.write(self.get_info_string())
+
     def scenario_variation_path(self, par_id, run_id):
         return os.path.join(
             self.get_env_outputfolder_path(),
@@ -360,6 +363,13 @@ class AbstractEnvironmentManager(object):
         numbered_scenario_name = "_".join([digits_parameter_id, digits_run_id])
 
         return "".join([numbered_scenario_name, self.VADERE_SCENARIO_FILE_TYPE])
+
+    def get_info_string(self):
+
+        if self.info_string is None:
+            return f"No information provided from {type(self)}."
+        else:
+            return self.info_string
 
 
 class VadereEnvironmentManager(AbstractEnvironmentManager):
@@ -475,6 +485,32 @@ class CoupledEnvironmentManager(AbstractEnvironmentManager):
             raise RuntimeError(f"None or too many 'ini' files " "found in environment.")
         return sc_files[0]
 
+
+    @classmethod
+    def create_variation_env_from_info_file(
+            cls,
+            path_info_file
+    ):
+
+        d = {}
+        with open(path_info_file) as f:
+            for line in f:
+                (key, val) = line.split()
+                d[key] = val
+
+        env = cls.create_variation_env(
+            basis_scenario = d["basis_scenario"],
+            ini_scenario =d["ini_path"],
+            base_path =d["base_path"],
+            env_name =d["env_name"],
+            handle_existing = "write_in")
+
+        return env
+
+
+
+
+
     @classmethod
     def create_variation_env(
         cls,
@@ -484,6 +520,8 @@ class CoupledEnvironmentManager(AbstractEnvironmentManager):
         env_name=None,
         handle_existing="ask_user_replace",
     ):
+
+        cls.set_info_string(basis_scenario, ini_scenario, base_path, env_name)
 
         cls.basis_scenario_name = os.path.basename(basis_scenario)
         # Check if environment already exists
@@ -495,8 +533,10 @@ class CoupledEnvironmentManager(AbstractEnvironmentManager):
         ini_path = os.path.dirname(ini_scenario)
 
         new_path = os.path.join(path_output_folder, "additional_rover_files")
-        copytree(ini_path, new_path, ignore=include_patterns("*.py", "*.xml"))
-        removeEmptyFolders(new_path)
+
+        if os.path.exists(new_path) is False:
+            copytree(ini_path, new_path, ignore=include_patterns("*.py", "*.xml"))
+            removeEmptyFolders(new_path)
 
         # Add vadere basis scenario used for the variation (i.e. sampling)
         if isinstance(basis_scenario, str):  # assume that this is a path
@@ -560,12 +600,14 @@ class CoupledEnvironmentManager(AbstractEnvironmentManager):
                 outfile.write(s)
 
         # Create the folder where all output is stored
-        os.mkdir(
-            os.path.join(
+
+        path_output_folder_rover = os.path.join(
                 path_output_folder,
                 CoupledEnvironmentManager.simulation_runs_output_folder,
             )
-        )
+
+        if os.path.exists(path_output_folder_rover) is False:
+            os.mkdir(path_output_folder_rover)
 
         return cls(base_path, env_name)
 
@@ -627,3 +669,11 @@ class CoupledEnvironmentManager(AbstractEnvironmentManager):
     def get_simulation_directory(self, par_id, run_id):
         prefix = CoupledEnvironmentManager.simulation_runs_single_folder_name
         return f"{prefix}_{par_id}_{run_id}"
+
+    @classmethod
+    def set_info_string(cls, basis_scenario, ini_path, base_path, env_name):
+        info = f"basis_scenario {basis_scenario}\n"
+        info = info + f"ini_path {ini_path}\n"
+        info = info + f"base_path {base_path}\n"
+        info = info + f"env_name {env_name}\n"
+        cls.info_string = info
